--- src/app/settings/player/play.css
@@ -1,3 +1,0 @@
-.maxHeightVideo {
-    max-height: 80dvh;
-}
