+++ src/components/playlist/cards/nicoVideoCard.tsx
@@ -5,11 +5,11 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import CircularProgress from '@mui/material/CircularProgress';
 
 // Types
-import { Playlist } from '@/types/playlist';
+import { VideoAbout } from '@/types/db';
 import nicoImg from "@/utils/niconico/nicoimg";
 import { SiNiconico } from "react-icons/si";
 
-export default function niconicoVideoCard({ props , item , deleteLoading , deletePlaylist } : {props : {ytid:string , playlistId:string  } , item: Playlist , deleteLoading: Array<string> , deletePlaylist: (id: string) => void }) {
+export default function niconicoVideoCard({ props , item , deleteLoading , deletePlaylist } : {props : {ytid:string , playlistId:string  } , item: VideoAbout , deleteLoading: Array<string> , deletePlaylist: (id: string) => void }) {
     return (
         <div
             className={`relative my-6 break-all sm:flex items-start gap-4 cursor-pointer rounded-lg shadow-md hover:bg-gray-100 transition-colors ${
