--- src/types/channel.ts
@@ -1,11 +1,0 @@
-export type Channel = { 
-    channelId: string, 
-    channelContent: { 
-        title: string, 
-        thumbnails: { 
-            medium: { 
-                url: string 
-            } 
-        } 
-    } 
-}
