--- src/types/history.ts
@@ -1,11 +1,0 @@
-export type History = { videoId: string, 
-    videoContent: {
-        title: string, 
-        channelTitle: string, 
-        thumbnail: { 
-            url: string, 
-            middleUrl: string | null, 
-            largeUrl: string | null 
-        } 
-    } 
-}
